changed bigint

+++ migrations/20260520162150_create_training.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS training
     trainer_id      TEXT      NOT NULL,
     available_from  TIMESTAMP NOT NULL,
     available_to    TIMESTAMP NOT NULL,
-    duration int NOT NULL,
+    duration BIGINT NOT NULL,
     additional_info TEXT      NOT NULL,
     PRIMARY KEY (id)
 );
